# be/src/main/resources/openapi/alumni-api.yaml
openapi: 3.0.3
info:
  title: Alumni Registry Management API
  description: API for handling the REST endpoints
  version: 1.1.0
  contact:
    name: Support Team
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Development server

paths:
  /alumni:
    post:
      summary: Create a new alumna/alumnus
      description: Create a record of a new alumna/alumnus within the system
      operationId: createAlumni
      tags:
        - Alumni
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlumniRequest'
      responses:
        '201':
          description: alumna/alumnus created with success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlumniResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Address:
      type: object
      required:
        - street
        - number
        - country
      properties:
        street:
          type: string
          description: Street name
          example: "streetname"
        number:
          type: string
          description: Street number
          example: "22"
        country:
          type: string
          description: Country name
          example: "country"

    Degree:
      type: object
      required:
        - university
        - year
      properties:
        university:
          type: string
          description: University name
          example: "Politecnico Milano"
        year:
          type: integer
          description: Year of completion
          example: 2004

    Education:
      type: object
      properties:
        master:
          $ref: '#/components/schemas/Degree'
        phd:
          $ref: '#/components/schemas/Degree'

    AlumniRequest:
      type: object
      required:
        - name
        - addresses
        - education
      properties:
        name:
          type: string
          description: Name
          example: "Mario Rossi"
          minLength: 1
          maxLength: 100
        addresses:
          type: array
          description: List of addresses
          items:
            $ref: '#/components/schemas/Address'
          minItems: 1
        education:
          $ref: '#/components/schemas/Education'

    AlumniResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of alumna/alumnus
          example: 1
        name:
          type: string
          description: Name of alumna/alumnus
          example: "Mario Rossi"
        addresses:
          type: array
          description: List of addresses
          items:
            $ref: '#/components/schemas/Address'
        education:
          $ref: '#/components/schemas/Education'
        createdAt:
          type: string
          format: date-time
          description: Timestamp of record creation
          example: "2024-02-02T10:30:00Z"

    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: "INVALID_REQUEST"
        message:
          type: string
          description: Error description
          example: "The name field is required"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-02-02T10:30:00Z"